Separated predicates in a separate module
